+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
  - DOI filter not working properly.
 
+### Added 
+- bioRxiv and medRxiv preprint servers.
+
 ## 1.0.40.10 - 2023-07-05
 ### Changed
 - [351] (https://github.com/CCSDForge/episciences/issues/351): formatted files available on the copy editing section